requirements.txt
robotframework==6.1.1
requests==2.32.2
robotframework-jsonlibrary==0.5
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-seleniumlibrary==6.7.1
selenium==4.31.0
mailosaur==7.19.0
